pyproject.toml
[tool.poetry]
name = "djangoauth"
version = "0.1.0"
description = ""
authors = ["ratchaphon1412 <ratchaphon.h111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.5"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
kafka-python = "^2.0.2"
lz4 = "^4.3.2"
crc32c = "^2.3.post0"
gunicorn = "^21.2.0"
django-seed = "^0.3.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
